fix(BLE-iOS): avoid null exception in case of GATT read errors
If a characteristics is encrypted, the central and peripheral must be paired in order for the central to access the data.
Depending on the BLE device settings, iOS shows a popup for pairing.
If the user cancels the pairing request or the request times out, the Value of the characteristic will be null and a non-null Error will be provided.