Extended class BluetoothDeviceInfo with SerialNumber and ProtocolStrings (iOS)
An ExternalAccessory (iOS) has various properties, including frequently used properties such as the device name or the device address. However, it can be useful for an application to also be able to read the serial number and protocol strings, e.g. to be able to filter a list of devices.
As our application only requires these properties on iOS devices, this functionality has only been implemented for iOS. When querying the protocol strings of other platforms, a PlatformNotSupportedException is thrown and an empty string is returned for the serial number.